backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.105.0
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.27.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.0
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.1
overrides==7.7.0
packaging==25.0
Pillow==10.4.0
pinecone-client==4.1.2
posthog==5.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytesseract==0.3.13
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.21.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0